DkStdRestApis support Dune+DkCoder; +2 libraries
ext/DkStdRestApis is externally pushed as open-source, so Dune is desirable. And the Dune bits let us check with compile-us.sh without having to do a time consuming ci/{build,distribute}-*.sh build.

+ Move ServerTiny into its own library

So tiny-httpd does not need to be in the dependency cone for DkStdRestApis_NotStripe. That also means that we re-introduce a [open DkStdRestApis_NotStripe] statement in the new DkStdRestApis_NotStripe_Tiny library.

+ Move Curl2 and CurlCmd into its own library

+ Change to new Curl API

+bugfixes: Correct module open bugs shown by Dune build